remove duplicate implementations in macros
