+++ src/derive/field.rs
@@ -49,7 +49,74 @@ macro_rules! field_common {
             /// Converts from an integer represented in little endian
             /// into its (congruent) `$field` representation.
             pub const fn from_raw(val: [u64; 4]) -> Self {
-                (&$field(val)).mul(&$r2)
+                #[cfg(feature = "asm")]
+                {
+                    let (r0, carry) = mac(0, val[0], $r2.0[0], 0);
+                    let (r1, carry) = mac(0, val[0], $r2.0[1], carry);
+                    let (r2, carry) = mac(0, val[0], $r2.0[2], carry);
+                    let (r3, r4) = mac(0, val[0], $r2.0[3], carry);
+
+                    let (r1, carry) = mac(r1, val[1], $r2.0[0], 0);
+                    let (r2, carry) = mac(r2, val[1], $r2.0[1], carry);
+                    let (r3, carry) = mac(r3, val[1], $r2.0[2], carry);
+                    let (r4, r5) = mac(r4, val[1], $r2.0[3], carry);
+
+                    let (r2, carry) = mac(r2, val[2], $r2.0[0], 0);
+                    let (r3, carry) = mac(r3, val[2], $r2.0[1], carry);
+                    let (r4, carry) = mac(r4, val[2], $r2.0[2], carry);
+                    let (r5, r6) = mac(r5, val[2], $r2.0[3], carry);
+
+                    let (r3, carry) = mac(r3, val[3], $r2.0[0], 0);
+                    let (r4, carry) = mac(r4, val[3], $r2.0[1], carry);
+                    let (r5, carry) = mac(r5, val[3], $r2.0[2], carry);
+                    let (r6, r7) = mac(r6, val[3], $r2.0[3], carry);
+
+                    // Montgomery reduction (first part)
+                    let k = r0.wrapping_mul($inv);
+                    let (_, carry) = mac(r0, k, $modulus.0[0], 0);
+                    let (r1, carry) = mac(r1, k, $modulus.0[1], carry);
+                    let (r2, carry) = mac(r2, k, $modulus.0[2], carry);
+                    let (r3, carry) = mac(r3, k, $modulus.0[3], carry);
+                    let (r4, carry2) = adc(r4, 0, carry);
+
+                    let k = r1.wrapping_mul($inv);
+                    let (_, carry) = mac(r1, k, $modulus.0[0], 0);
+                    let (r2, carry) = mac(r2, k, $modulus.0[1], carry);
+                    let (r3, carry) = mac(r3, k, $modulus.0[2], carry);
+                    let (r4, carry) = mac(r4, k, $modulus.0[3], carry);
+                    let (r5, carry2) = adc(r5, carry2, carry);
+
+                    let k = r2.wrapping_mul($inv);
+                    let (_, carry) = mac(r2, k, $modulus.0[0], 0);
+                    let (r3, carry) = mac(r3, k, $modulus.0[1], carry);
+                    let (r4, carry) = mac(r4, k, $modulus.0[2], carry);
+                    let (r5, carry) = mac(r5, k, $modulus.0[3], carry);
+                    let (r6, carry2) = adc(r6, carry2, carry);
+
+                    let k = r3.wrapping_mul($inv);
+                    let (_, carry) = mac(r3, k, $modulus.0[0], 0);
+                    let (r4, carry) = mac(r4, k, $modulus.0[1], carry);
+                    let (r5, carry) = mac(r5, k, $modulus.0[2], carry);
+                    let (r6, carry) = mac(r6, k, $modulus.0[3], carry);
+                    let (r7, _) = adc(r7, carry2, carry);
+
+                    // Montgomery reduction (sub part)
+                    let (d0, borrow) = sbb(r4, $modulus.0[0], 0);
+                    let (d1, borrow) = sbb(r5, $modulus.0[1], borrow);
+                    let (d2, borrow) = sbb(r6, $modulus.0[2], borrow);
+                    let (d3, borrow) = sbb(r7, $modulus.0[3], borrow);
+
+                    let (d0, carry) = adc(d0, $modulus.0[0] & borrow, 0);
+                    let (d1, carry) = adc(d1, $modulus.0[1] & borrow, carry);
+                    let (d2, carry) = adc(d2, $modulus.0[2] & borrow, carry);
+                    let (d3, _) = adc(d3, $modulus.0[3] & borrow, carry);
+
+                    $field([d0, d1, d2, d3])
+                }
+                #[cfg(not(feature = "asm"))]
+                {
+                    (&$field(val)).mul(&$r2)
+                }
             }
 
             /// Attempts to convert a little-endian byte representation of
@@ -62,6 +129,16 @@ macro_rules! field_common {
             /// little-endian byte order.
             pub fn to_bytes(&self) -> [u8; 32] {
                 <Self as ff::PrimeField>::to_repr(self)
+            }
+
+            /// Lexicographic comparison of Montgomery forms.
+            #[inline(always)]
+            const fn is_less_than(x: &[u64; 4], y: &[u64; 4]) -> bool {
+                let (_, borrow) = sbb(x[0], y[0], 0);
+                let (_, borrow) = sbb(x[1], y[1], borrow);
+                let (_, borrow) = sbb(x[2], y[2], borrow);
+                let (_, borrow) = sbb(x[3], y[3], borrow);
+                borrow >> 63 == 1
             }
         }
 
@@ -411,16 +488,6 @@ macro_rules! field_arithmetic {
                     (((self.0[0] | self.0[1] | self.0[2] | self.0[3]) == 0) as u64).wrapping_sub(1);
 
                 $field([d0 & mask, d1 & mask, d2 & mask, d3 & mask])
-            }
-
-            /// Lexicographic comparison of Montgomery forms.
-            #[inline(always)]
-            fn is_less_than(x: &[u64; 4], y: &[u64; 4]) -> bool {
-                let (_, borrow) = sbb(x[0], y[0], 0);
-                let (_, borrow) = sbb(x[1], y[1], borrow);
-                let (_, borrow) = sbb(x[2], y[2], borrow);
-                let (_, borrow) = sbb(x[3], y[3], borrow);
-                borrow >> 63 == 1
             }
         }
     };
